+++ lib/feature/statistics/data/models/statistic_model.dart
@@ -185,7 +185,7 @@ class StatisticSummary {
     bfsDate: json["bfs_date"],
     uniqueValue: json["unique_value"],
     colorValue: json['color_value'] != null ? HexColor(json['color_value']) : AppColors.colors[Helper.index(7)],
-    sortValue: json['sort_value']== null ? 0 : int.parse(json['sort_value'])  
+    sortValue: json['sort_value']== null ? 0.0 : double.parse(json['sort_value'])
   );
 
   Map<String, dynamic> toJson() => {
